Optimise PRNG in src/rsrca.c ...
- Run cpp_mt19937 in a child process
  - saw performance improvement by changing to 64 bit variant PRNG
  - to have some level of parallelism
  - measure CPU time spent in PRNG

Dry run throughput improved. However, it was proven that the major
bottleneck is from the raw socket IO(sendto()), not the PRNG. Also,
there's additional syscall cost introduced(write() and read() to and
from the pipe), but given that the the child process hardly comsumes CPU
time in "wet run", it was determined that the overhead from context
switching is negligible.